refactor: extract shared FIDE Article 5.2 color rules into utilities
move rankPreference, FIDE_COLOR_RULES and ROUND_1_COLOR_RULE to
utilities.ts — Dutch/Lim use FIDE_COLOR_RULES directly, Dubov/Burstein
prepend ROUND_1_COLOR_RULE. removes ~220 lines of duplicated color
allocation logic across four pairing systems.
