+++ src/dubov.ts
@@ -19,6 +19,8 @@
 
 import { maxWeightMatching } from './blossom.js';
 import {
+  FIDE_COLOR_RULES,
+  ROUND_1_COLOR_RULE,
   allocateColor,
   assignBye,
   buildPlayerStates,
@@ -29,7 +31,7 @@ import { buildEdgeWeight } from './weights.js';
 import type { DynamicUint } from './dynamic-uint.js';
 import type { PairOptions, TraceCallback } from './trace.js';
 import type { Game, PairingResult, Player } from './types.js';
-import type { ColorRule, PlayerState } from './utilities.js';
+import type { PlayerState } from './utilities.js';
 import type { BracketContext, Criterion } from './weights.js';
 
 // ---------------------------------------------------------------------------
@@ -62,81 +64,7 @@ function computeARO(state: PlayerState, players: Player[]): number {
 // FIDE Article 5 colour rules
 // ---------------------------------------------------------------------------
 
-function rankPreference(s: PlayerState['preferenceStrength']): number {
-  if (s === 'absolute') return 3;
-  if (s === 'strong') return 2;
-  if (s === 'mild') return 1;
-  return 0;
-}
-
-const DUBOV_COLOR_RULES: ColorRule[] = [
-  // 5.2.1 (round 1): Both players have no history — odd TPN gets initial colour (white)
-  (hrp, opp) => {
-    const hrpHasHistory = hrp.colorHistory.some((c) => c !== undefined);
-    const oppHasHistory = opp.colorHistory.some((c) => c !== undefined);
-    if (!hrpHasHistory && !oppHasHistory) {
-      return hrp.tpn % 2 === 1 ? 'hrp-white' : 'hrp-black';
-    }
-    return 'continue';
-  },
-  // 5.2.2 Grant both colour preferences (if they differ)
-  (hrp, opp) => {
-    if (
-      hrp.preferredColor !== undefined &&
-      opp.preferredColor !== undefined &&
-      hrp.preferredColor !== opp.preferredColor
-    ) {
-      return hrp.preferredColor === 'white' ? 'hrp-white' : 'hrp-black';
-    }
-    return 'continue';
-  },
-  // 5.2.3 Grant stronger preference; both absolute → wider colorDiff wins
-  (hrp, opp) => {
-    const hrpS = rankPreference(hrp.preferenceStrength);
-    const oppS = rankPreference(opp.preferenceStrength);
-
-    if (hrpS > oppS && hrp.preferredColor !== undefined) {
-      return hrp.preferredColor === 'white' ? 'hrp-white' : 'hrp-black';
-    }
-    if (oppS > hrpS && opp.preferredColor !== undefined) {
-      return opp.preferredColor === 'white' ? 'hrp-black' : 'hrp-white';
-    }
-    // Both absolute: wider colorDiff wins
-    if (hrpS === 3 && oppS === 3) {
-      const hrpAbs = Math.abs(hrp.colorDiff);
-      const oppAbs = Math.abs(opp.colorDiff);
-      if (hrpAbs > oppAbs && hrp.preferredColor !== undefined) {
-        return hrp.preferredColor === 'white' ? 'hrp-white' : 'hrp-black';
-      }
-      if (oppAbs > hrpAbs && opp.preferredColor !== undefined) {
-        return opp.preferredColor === 'white' ? 'hrp-black' : 'hrp-white';
-      }
-    }
-    return 'continue';
-  },
-  // 5.2.4 Alternate from most recent divergent round
-  (hrp, opp) => {
-    const minLength = Math.min(
-      hrp.colorHistory.length,
-      opp.colorHistory.length,
-    );
-    for (let index = minLength - 1; index >= 0; index--) {
-      const h = hrp.colorHistory[index];
-      const o = opp.colorHistory[index];
-      if (h !== undefined && o !== undefined && h !== o) {
-        return h === 'white' ? 'hrp-black' : 'hrp-white';
-      }
-    }
-    return 'continue';
-  },
-  // 5.2.5 Grant HRP's preference
-  (hrp) => {
-    if (hrp.preferredColor !== undefined) {
-      return hrp.preferredColor === 'white' ? 'hrp-white' : 'hrp-black';
-    }
-    return 'continue';
-  },
-];
+const DUBOV_COLOR_RULES = [ROUND_1_COLOR_RULE, ...FIDE_COLOR_RULES];
 
 // Rank comparator for allocateColor: lower TPN = higher rank in Dubov
 function dubovRankCompare(a: PlayerState, b: PlayerState): number {
