+++ src/lim.ts
@@ -25,6 +25,7 @@
 
 import { maxWeightMatching } from './blossom.js';
 import {
+  FIDE_COLOR_RULES,
   allocateColor,
   assignBye,
   buildPlayerStates,
@@ -35,81 +36,8 @@ import { buildEdgeWeight } from './weights.js';
 import type { DynamicUint } from './dynamic-uint.js';
 import type { PairOptions, TraceCallback } from './trace.js';
 import type { Game, PairingResult, Player } from './types.js';
-import type { ColorRule, PlayerState } from './utilities.js';
+import type { PlayerState } from './utilities.js';
 import type { BracketContext, Criterion } from './weights.js';
-
-// ---------------------------------------------------------------------------
-// FIDE Article 5.2 colour rules (same as Dutch)
-// ---------------------------------------------------------------------------
-
-function rankPreference(s: PlayerState['preferenceStrength']): number {
-  if (s === 'absolute') return 3;
-  if (s === 'strong') return 2;
-  if (s === 'mild') return 1;
-  return 0;
-}
-
-const LIM_COLOR_RULES: ColorRule[] = [
-  // 5.2.1 Grant both colour preferences (if they differ)
-  (hrp, opp) => {
-    if (
-      hrp.preferredColor !== undefined &&
-      opp.preferredColor !== undefined &&
-      hrp.preferredColor !== opp.preferredColor
-    ) {
-      return hrp.preferredColor === 'white' ? 'hrp-white' : 'hrp-black';
-    }
-    return 'continue';
-  },
-  // 5.2.2 Grant stronger preference; both absolute → wider colorDiff wins
-  (hrp, opp) => {
-    const hrpS = rankPreference(hrp.preferenceStrength);
-    const oppS = rankPreference(opp.preferenceStrength);
-
-    if (hrpS > oppS && hrp.preferredColor !== undefined) {
-      return hrp.preferredColor === 'white' ? 'hrp-white' : 'hrp-black';
-    }
-    if (oppS > hrpS && opp.preferredColor !== undefined) {
-      return opp.preferredColor === 'white' ? 'hrp-black' : 'hrp-white';
-    }
-    // Both absolute: wider colorDiff wins
-    if (hrpS === 3 && oppS === 3) {
-      const hrpAbs = Math.abs(hrp.colorDiff);
-      const oppAbs = Math.abs(opp.colorDiff);
-      if (hrpAbs > oppAbs && hrp.preferredColor !== undefined) {
-        return hrp.preferredColor === 'white' ? 'hrp-white' : 'hrp-black';
-      }
-      if (oppAbs > hrpAbs && opp.preferredColor !== undefined) {
-        return opp.preferredColor === 'white' ? 'hrp-black' : 'hrp-white';
-      }
-    }
-    return 'continue';
-  },
-  // 5.2.3 Alternate from most recent divergent round
-  (hrp, opp) => {
-    const minLength = Math.min(
-      hrp.colorHistory.length,
-      opp.colorHistory.length,
-    );
-    for (let index = minLength - 1; index >= 0; index--) {
-      const h = hrp.colorHistory[index];
-      const o = opp.colorHistory[index];
-      if (h !== undefined && o !== undefined && h !== o) {
-        return h === 'white' ? 'hrp-black' : 'hrp-white';
-      }
-    }
-    return 'continue';
-  },
-  // 5.2.4 Grant HRP's preference
-  (hrp) => {
-    if (hrp.preferredColor !== undefined) {
-      return hrp.preferredColor === 'white' ? 'hrp-white' : 'hrp-black';
-    }
-    return 'continue';
-  },
-  // 5.2.5 Odd TPN → initial colour (white)
-  (hrp) => (hrp.tpn % 2 === 1 ? 'hrp-white' : 'hrp-black'),
-];
 
 // Rank comparator for allocateColor: lower TPN = higher rank
 function limRankCompare(a: PlayerState, b: PlayerState): number {
@@ -542,7 +470,7 @@ function pair(
   // Allocate colours
   // -------------------------------------------------------------------------
   const pairings = allPairedTuples.map(([a, b]) =>
-    allocateColor(a, b, LIM_COLOR_RULES, limRankCompare),
+    allocateColor(a, b, FIDE_COLOR_RULES, limRankCompare),
   );
 
   if (trace) {
